apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "dash>=4.1.0",
    "duckdb>=1.5.2",
    "edgartools>=5.30.0",
    "fastapi>=0.136.0",
    "google-adk>=1.31.0",
    "httpx>=0.28.1",
    "litellm>=1.83.0",
    "pandas>=3.0.2",
    "plotly[express]>=6.7.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.44.0",
    "yfinance>=1.3.0",
]
